portable image: the crates above bun_sys take the Fd and the Stat of the image [skip ci]
"Fd for both kinds of host" and "bun_sys has its code for Windows and
for POSIX" gave the portable image the Fd of Windows (a u64 that holds a
HANDLE or an int) and a Stat with the fields of uv_stat_t, and adapted
bun_core, bun_errno, bun_paths and bun_sys. The crates above them were
never compiled with --cfg=bun_portable after that, and all of bun as an
image needs them: bun_uws_sys, bun_spawn_sys, bun_io, bun_resolver,
bun_install, bun_jsc, bun_runtime.

As bun_sys does it:
- Fd::native() where the argument does not say which kind is meant (a
  variadic syscall, a cast, a formatted string): Fd::posix() in code of
  a POSIX system, and a `let` of the type FdNative where a build for
  Windows compiles the same line.
- A field of Stat that goes to the C library is cast to the type of the
  C library (mode_t, uid_t). The cast changes nothing where the field
  has that type.
- The times of a Stat are `mtim` in the image, as in a build for
  Windows: the two `cfg` of each place name bun_portable.

Blob: a file descriptor that is read from a serialized record is
refused when its int is negative. The check reads the int out of the
u64 in the image, on a host that has file descriptors.

No line of a build for one OS changes its meaning.

+++ src/install/PackageInstaller.rs
@@ -2142,9 +2142,9 @@ impl<'a> PackageInstaller<'a> {
                                 // no preconditions), so no `unsafe` needed.
                                 // `st_mode` is u16 on FreeBSD, u32 elsewhere; widen.
                                 let st_mode = stat.st_mode as u32;
-                                let is_writable = if stat.st_uid == bun_sys::c::getuid() {
+                                let is_writable = if stat.st_uid as u32 == bun_sys::c::getuid() {
                                     st_mode & bun_sys::S::IWUSR > 0
-                                } else if stat.st_gid == bun_sys::c::getgid() {
+                                } else if stat.st_gid as u32 == bun_sys::c::getgid() {
                                     st_mode & bun_sys::S::IWGRP > 0
                                 } else {
                                     st_mode & bun_sys::S::IWOTH > 0
